fixing windows warnings
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@86357 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: b8398ea47fc557db4afbea8646ed2276e7e26c0d

+++ Rec/Tr/TrackTools/src/VeloCaloBuilder.h
@@ -69,10 +69,10 @@ class VeloCaloBuilder : public GaudiTupleAlg {
   ICaloGetterTool* getter;
 
   // -- variables from HltVeloEcalMatch
-  double 	m_ptkickConstant;
-  double 	m_zKick;
-  double 	m_eres [2];
-  double 	m_eCorrect;
+  float 	m_ptkickConstant;
+  float 	m_zKick;
+  float 	m_eres [2];
+  float 	m_eCorrect;
 
 };
 #endif
